second

+++ api.js
@@ -10,17 +10,39 @@ router.get("/", (req, res) => {
   res.send("Hello from the router!");
 });
 
-router.get("/users", (req, res) => {
-  User.find({})
-    .then((users) => {
-      res.send(users);
-    })
-    .catch((err) => {
-      res.status(400).send(err);
-    });
-});
+// router.get("/users/find", (req, res) => {
+//   User.aggregate([
+//     {
+//       $group: {
+//         _id: { name: "$name" },
+//         count: { $sum: 1 },
+//       },
+//     },
+//     {
+//       $match: {
+//         count: { $gt: 1 },
+//       },
+//     },
+//   ])
+//     .then((duplicateNames) => {
+//       res.send(duplicateNames);
+//     })
+//     .catch((err) => {
+//       res.status(400).send(err);
+//     });
+// });
 
-router.get("/users/find", (req, res) => {
+// router.get("/users", (req, res) => {
+//   User.find({})
+//     .then((users) => {
+//       res.send(users);
+//     })
+//     .catch((err) => {
+//       res.status(400).send(err);
+//     });
+// });
+
+router.get("/users", (req, res) => {
   User.find(req.query)
     .then((users) => {
       res.send(users);
@@ -30,24 +52,59 @@ router.get("/users/find", (req, res) => {
     });
 });
 
+// router.get("/users", (req, res) => {
+//   // If the 'name' query parameter is present, use it to filter by name
+//   const nameQuery = req.query.name ? { name: req.query.name } : {};
+
+//   User.find(nameQuery)
+//     .then((users) => {
+//       res.send(users);
+//     })
+//     .catch((err) => {
+//       res.status(400).send(err);
+//     });
+// });
+
+router.delete("/users/:id", (req, res) => {
+  if(ObjectId.isValid(req.params.id) === false) {
+    res.status(400).send("Invalid ID");
+    return
+  }
+  User.findByIdAndDelete(req.params.id)
+  .then((user) => {
+    res.send(user);
+  })
+  .catch((err) => {
+    res.status(400).send(err);
+  });
+})
+
 router.get("/users/tasks/:id", (req, res) => {
+  if(ObjectId.isValid(req.params.id) === false) {
+    res.status(400).send("Invalid ID");
+    return
+  }
   Task.find({ assignedTo: req.params.id })
-    .then((user) => {
-      res.send(user);
-    })
-    .catch((err) => {
-      res.status(400).send(err);
-    });
+  .then((user) => {
+    res.send(user);
+  })
+  .catch((err) => {
+    res.status(400).send(err);
+  });
 });
 
 router.get("/users/:id", (req, res) => {
+  if(ObjectId.isValid(req.params.id) === false) {
+    res.status(400).send("Invalid ID");
+    return
+  }
   User.findById(req.params.id)
-    .then((user) => {
-      res.send(user);
-    })
-    .catch((err) => {
-      res.status(400).send(err);
-    });
+  .then((user) => {
+    res.send(user);
+  })
+  .catch((err) => {
+    res.status(400).send(err);
+  });
 });
 
 router.post("/users", (req, res) => {
@@ -61,25 +118,40 @@ router.post("/users", (req, res) => {
   if (!req.body.name) {
     res.status(400).send("Name is missing");
   }
+
+  User.findOne({ name: req.body.name })
+  .then((user) => {
+    if (user) {
+      res.status(400).send("User already exists");
+    }
+  })
+  .catch((err) => {
+    res.status(400).send(err);
+  });
+
   const user = new User(req.body);
   user
-    .save()
-    .then((user) => {
-      res.send(user);
-    })
-    .catch((err) => {
-      res.status(400).send(err);
-    });
+  .save()
+  .then((user) => {
+    res.send(user);
+  })
+  .catch((err) => {
+    res.status(400).send(err);
+  });
 });
 
 router.delete("/users/:id", (req, res) => {
+  if(ObjectId.isValid(req.params.id) === false) {
+    res.status(400).send("Invalid ID");
+    return
+  }
   User.findByIdAndDelete(req.params.id)
-    .then((user) => {
-      res.send(user);
-    })
-    .catch((err) => {
-      res.status(400).send(err);
-    });
+  .then((user) => {
+    res.send(user);
+  })
+  .catch((err) => {
+    res.status(400).send(err);
+  });
 });
 
 router.post("/tasks", (req, res) => {
@@ -96,12 +168,12 @@ router.post("/tasks", (req, res) => {
     !req.body.status ||
     !["pending", "working", "review", "done", "archive"].includes(
       req.body.status
-    )
-  ) {
-    res.status(400).send("Status is missing or invalid");
-  }
-  const task = new Task(req.body);
-  task
+      )
+    ) {
+      res.status(400).send("Status is missing or invalid");
+    }
+    const task = new Task(req.body);
+    task
     .save()
     .then((task) => {
       res.send(task);
@@ -121,26 +193,34 @@ router.get("/tasks", (req, res) => {
     sortOptions.updatedAt = -1;
   }
   Task.find(query)
-    .sort(sortOptions)
-    .then((tasks) => {
-      res.send(tasks);
-    })
-    .catch((err) => {
-      res.status(400).send(err);
-    });
+  .sort(sortOptions)
+  .then((tasks) => {
+    res.send(tasks);
+  })
+  .catch((err) => {
+    res.status(400).send(err);
+  });
 });
 
 router.get("/tasks/:id", (req, res) => {
+  if(ObjectId.isValid(req.params.id) === false) {
+    res.status(400).send("Invalid ID");
+    return
+  }
   Task.findById(req.params.id)
-    .then((task) => {
-      res.send(task);
-    })
-    .catch((err) => {
-      res.status(400).send(err);
-    });
+  .then((task) => {
+    res.send(task);
+  })
+  .catch((err) => {
+    res.status(400).send(err);
+  });
 });
 
 router.put("/tasks/:id", async (req, res) => {
+  if(ObjectId.isValid(req.params.id) === false) {
+    res.status(400).send("Invalid ID");
+    return
+  }
   let { assignedTo, status } = req.body;
   if (!assignedTo) {
     assignedTo = null;
@@ -149,11 +229,12 @@ router.put("/tasks/:id", async (req, res) => {
   }
 
   if (status) {
-    const task = await Task.findById(req.params.id, { status }, { new: true })
-      .then((task) => task)
-      .catch((err) => {
-        res.status(400).send(err);
-      });
+    const task = await Task.findById(req.params.id)
+    .then((task) => task)
+    .catch((err) => {
+      res.status(400).send(err);
+    });
+    console.log(task)
     if (task.status === "done" && status !== "archive") {
       res.status(400).send("Can't change task status from done to " + status);
     }
@@ -169,26 +250,30 @@ router.put("/tasks/:id", async (req, res) => {
   console.log(updatedDoc);
 
   Task.findByIdAndUpdate(req.params.id, updatedDoc, { new: true })
-    .then((task) => {
-      res.send(task);
-    })
-    .catch((err) => {
-      res.status(400).send(err);
-    });
+  .then((task) => {
+    res.send(task);
+  })
+  .catch((err) => {
+    res.status(400).send(err);
+  });
 });
 
 router.delete("/tasks/:id", async (req, res) => {
+  if(ObjectId.isValid(req.params.id) === false) {
+    res.status(400).send("Invalid ID");
+    return
+  }
   Task.findByIdAndDelete(
     req.params.id,
     { $set: { isDeleted: true } },
     { new: true }
-  )
+    )
     .then((task) => {
       res.send(task);
     })
     .catch((err) => {
       res.status(400).send(err);
     });
-});
+  });
 
-module.exports = router;
+  module.exports = router;
